+++ src/gpucb.h
@@ -5,42 +5,42 @@
 
 #include <stdbool.h>
 
-double GRID_MIN_;
-double GRID_INC_;
-double BETA_;
+float GRID_MIN_;
+float GRID_INC_;
+float BETA_;
 extern const char *tag[10];
 // Allocate memory
-double *T_;
+float *T_;
 int *X_;
-double *X_grid_;
+float *X_grid_;
 bool *sampled_;
-double *mu_;
-double *sigma_;
-double *K_;
-double *L_;
+float *mu_;
+float *sigma_;
+float *K_;
+float *L_;
 
 int I_;
 int N_;
 
-double function_baseline(double x, double y);
+float function_baseline(float x, float y);
 
-void learn_baseline(double *X_grid,
-                    double *K,
-                    double *L_T,
+void learn_baseline(float *X_grid,
+                    float *K,
+                    float *L_T,
                     bool *sampled,
                     int *X,
-                    double *T,
+                    float *T,
                     int t,
                     int maxIter,
-                    double *mu,
-                    double *sigma,
-                    double(*kernel)(double *, double *, double *, double *),
-                    const double beta,
+                    float *mu,
+                    float *sigma,
+                    float(*kernel)(float *, float *, float *, float *),
+                    const float beta,
                     int n);
 
-double kernel2_baseline(double *x1, double *y1, double *x2, double *y2);
+float kernel2_baseline(float *x1, float *y1, float *x2, float *y2);
 
-void initialize_meshgrid_baseline(double *X_grid, int n, double min, double inc);
+void initialize_meshgrid_baseline(float *X_grid, int n, float min, float inc);
 
 void initialize(const int, const int);
 
@@ -48,28 +48,28 @@ void run();
 
 void clean();
 
-int gpucb_baseline(int maxIter, int n, double grid_min, double grid_inc);
+int gpucb_baseline(int maxIter, int n, float grid_min, float grid_inc);
 
-void cholesky_baseline(double *A, int n, int size);
+void cholesky_baseline(float *A, int n, int size);
 
-void incremental_cholesky_baseline(double *A, double *A_T, int n1, int n2, int size);
+void incremental_cholesky_baseline(float *A, float *A_T, int n1, int n2, int size);
 
-void cholesky_solve2_baseline(int d, int size, double *LU, double *b, double *x, int lower);
+void cholesky_solve2_baseline(int d, int size, float *LU, float *b, float *x, int lower);
 
-void cholesky_solve_baseline(int d, double *LU, double *b, double *x);
+void cholesky_solve_baseline(int d, float *LU, float *b, float *x);
 
-void transpose_baseline(double *M, double *M_T, int d, int size);
+void transpose_baseline(float *M, float *M_T, int d, int size);
 
-void gp_regression_baseline(double *X_grid,
-                            double *K,
-                            double *L_T,
+void gp_regression_baseline(float *X_grid,
+                            float *K,
+                            float *L_T,
                             int *X,
-                            double *T,
+                            float *T,
                             int t,
                             int maxIter,
-                            double   (*kernel)(double *, double *, double *, double *),
-                            double *mu,
-                            double *sigma,
+                            float   (*kernel)(float *, float *, float *, float *),
+                            float *mu,
+                            float *sigma,
                             int n);
 
 #endif //FASTCODE_GPUCB_GPUCB_H
